requirements.txt
Django==2.0
psycopg2-binary
djangorestframework
gunicorn
django-heroku
pytest-django
